Fix: 통행불편 제보쪽 코드 리뷰 개선 코드 반영(#24)

+++ src/main/java/org/sesac/slopedbe/roadreport/service/RoadReportCenterService.java
@@ -1,5 +1,6 @@
 package org.sesac.slopedbe.roadreport.service;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -8,5 +9,5 @@ import org.sesac.slopedbe.roadreport.model.entity.RoadReportCenter;
 
 public interface RoadReportCenterService {
 	List<RoadReportCenterDTO> getComplaintCenterList();
-	Optional<RoadReportCenter> findClosestCenter(Double longitude, Double latitude, String mappingCity);
+	Optional<RoadReportCenter> findClosestCenter(BigDecimal longitude, BigDecimal latitude, String mappingCity);
 }
